Show stdout even if stderr contains data.
Output on stderr unfortunately does not necessarily
mean something went wrong.
When the file was encrypted to multiple keys, one of
them was on a SmartCard, and that SmartCard was removed,
the following printout appears on stderr:
"gpg: selecting openpgp failed: ec=6.108"
Decryption still succeeds if another key is available.

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>