Add WebDAV support.
Not added to the config dialog yet.
Note: The Debian version of fusedav will not work with several
WebDAV implementations.
All necessary patches are on the Debian bugtracker though (as
upstream is dead and the "official" successor project on github
doesn't seem to work at all with any WebDAV implementation I
could try).

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>