Modified the clipboard logic to allow for on-demand copy to clipboard.
Changed the clipboard config entry to be a combobox instead of a checkbox.  It has 3 states:
Never copy to the clipboard
Always copy to the clipboard
On-demand copy to the clipboard

Added a 'Password' button to the main window which copies the password to the clipboard when pressed.
The button is disabled when there is no password loaded or passwords are set to never copy to the clipboard.

Added a context menu entry to the text browser to copy the password to the clipboard.

Moved some of the clipboard code around to make it more flexible.  Clippedpass is always set when the password could be copied.