Fix & make clearClipboard more robust
The current version had 2 bugs:
1) The clippedText was never cleared, that isn't that much of a problem
since it is overriden anyways
2) The condition to check if clearing should happen was always true
since it compared the clippedText member variable with itself. The
clippedText variables that are create above the condition are out of
scope.

I fixed these bugs and clear both clipboard types if they contain the
given text, ignoring the setting, which should make it more robust.