Use FileContent parsing in MainWindow, contributes to #107
This also changes the behavior of the main window:
If content should be hidden nothing is shown.
Else the password is always shown as a field (either with hidden with
dots or openly visibly.) the rest is shown as plain text or as field
depending on the template settings.

This is results in a nicer ui and needs one translation (Password
hidden) less.