Correct and improve Dutch translations
The Dutch translations contained some errors and some translations
that deviated substantially from the original. This change corrects
errors and makes some changes. The focus was on keeping the
translations close to the original and use terminology consistently.
(In case the originals are actually worse descriptions, it is those
that should be changed first, before it is propagated to the
translations.)