Fix character encoding issues for non-UTF-8 locales.
Since we (most sensibly) encode text as UTF-8 before
encrypting we should assume that the password files
contain UTF-8 when decrypting, instead of the current
locale encoding.
This is the biggest issue on Windows, since it doesn't
even officially support locales with UTF-8 encoding.
For compatibility, detect if the data is not valid UTF-8
and fall back to Qt's BOM based approach, which provides
support for UTF-16 and falls back to current locale encoding.

Fixes issue #412