Fix missing app ID and icon on Wayland.
X11/Wayland needs to know the name of the .desktop file to show a dock
icon and application name. X11 has various means of guessing the
filename (often WM_NAME). Wayland is a bit more strict, and requires
that either the filename match the AppId or the .desktop filename be
specified.