Don't connect MainWindow::passShowHandlerFinished signal two times
QtPass::connectPassSignalHandlers() will be called twice, the first time
for the real pass and the second time for a pass imitator.
This means that we can't connect MainWindow::passShowHandlerFinished signal
in this function or it will be connected (and so then invoked) twice.

Connect it in the QtPass::connectPassSignalHandlers() single caller
instead.