Emit passShowHandlerFinished signal in MainWindow::passShowHandler()
Commit 3cb140ca697367 ("Trying to use QtPass as process handler and connector for Pass")
removed a DisplayInTextBrowser() call from MainWindow::passShowHandler()
and added a similar QtPass::showInTextBrowser() function that is invoked by
QtPass::passShowHandlerFinished() slot.

This slot is in turn connected to MainWindow::passShowHandlerFinished
signal which wasn't emitted anywhere in the code.

Emit it where the old DisplayInTextBrowser() call was so password entry
non-template details are actually displayed upon selecting it in the main
QtPass window.