Don't show a TOTP secret when selecting a password entry in the main window
Knowing the TOTP secret for a password entry allows somebody to recreate
the whole OTP sequence so it definitely shouldn't be displayed in the
clear.

In fact, it shouldn't be displayed at all in the main window since the
proper way to utilize a TOTP entry is to click the "OTP" button to generate
a new OTP (rather than to copy the secret to the clipboard like it was a
password).

The password edit dialog isn't affected by this change and will still show
the whole entry, including its TOTP secret if present.