Make sure resources are built just once
Currently, the project resources are built twice: the first time in the
"main" directory and the second time in the "src" directory.

This is because they are specified both in main.pro and src.pro qmake
project files (and in qtpass.pro, too).

They should be specified in just one such project file instead so they
are built just once.

Let's leave them in src.pro since that's where localization *.ts -> *.qm
dependencies are provided.
For this, a Q_INIT_RESOURCE() statement is needed so they are properly
linked to by the main application.